Separated Modals into individual components, added bonus points for diagonal words
